Generate manpage via `clap_mangen`
