+++ Cargo.toml
@@ -30,3 +30,6 @@ criterion = { version = "0.3", features = ["html_reports"] }
 name = "find_secrets"
 harness = false
 
+[build-dependencies]
+clap = { version = "4.2.5", features = ["derive"] }
+clap_mangen = "0.2.15"
